+++ src/transfer/request.ts
@@ -9,7 +9,7 @@ import { ShowConfirmMessage } from "../modules/vscode";
 import { filePropertiesTree } from "../ui/explorer/filepropertiestree";
 import logger from "../ui/logger";
 import statusBar, { Icon } from "../ui/statusbar";
-import { DoesFileExist, Validate } from "./gate";
+import { DoesFileExist, DoesFolderExist, Validate } from "./gate";
 import path = require("path");
 
 
@@ -37,7 +37,29 @@ export async function DeleteFile(uri: Uri, userConfig: UserConfig, system: Syste
     statusBar.updateBar('Deleted', Icon.success, { duration: 1 })
 
 }
+export async function DeleteFolder(uri: Uri, userConfig: UserConfig, system: SystemConfig) {
+    if (!await Validate(userConfig, system, uri.fsPath)) {
+        return false;
+    }
+    const sourcePath = GetRemotePath(uri.fsPath, userConfig);
+    if (!await DoesFolderExist(sourcePath, system)) {
+        logger.error('Folder does not exist.');
+        return false;
+    }
 
+    if (!await ShowConfirmMessage('Are you sure you want to delete this folder from the server? You can\'t undo this action.\nFolder: ' + sourcePath)) {
+        return;
+    }
+
+    statusBar.updateBar('Deleting', Icon.spinLoading, { duration: -1 });
+    const response = await deleteBatchService.call({ host: system.host, port: system.port }, sourcePath);
+    if (response) {
+        const folderName = path.basename(sourcePath);
+        logger.infos("Delete Folder", folderName + ": " + response?.Rowsets?.Messages?.Message);
+    }
+    statusBar.updateBar('Deleted', Icon.success, { duration: 1 })
+
+}
 
 export async function GetFileProperties(uri: Uri, userConfig: UserConfig, system: SystemConfig) {
     if (!await ValidatePath(uri.fsPath, userConfig)) return null;