+++ src/ui/progresswindow.ts
@@ -1,0 +1,36 @@
+import { CancellationToken, ProgressLocation, window } from "vscode";
+
+export function CreateProgressWindow(title: string, onCancel?: () => void) {
+    const data: { percent: number, end: () => void, thenable: Thenable<void>, token: CancellationToken } = { percent: 0, end: null, token: null, thenable: null };
+
+    data.thenable = window.withProgress({
+        location: ProgressLocation.Window,
+        cancellable: true,
+        title
+    }, async (progress, token) => {
+        let end = false;
+        if (onCancel)
+            token.onCancellationRequested(onCancel);
+        data.token = token;
+        data.end = () => {
+            end = true;
+            progress.report({
+                increment: 100,
+                message: ' ' + 100 + ' %'
+            });
+        };
+
+        progress.report({ increment: 0 });
+
+        while (!end) {
+            progress.report({
+                increment: data.percent,
+                message: ' ' + data.percent + ' %'
+            })
+            await new Promise(r => setTimeout(r, 1000 / 30));
+        }
+        progress.report({ increment: 100 });
+    });
+
+    return data;
+}